refactor: use float64 rather than i8 for for VolumeAdjustments
The range needed to be expanded from -120 to +120 to -120 to +135, so
the options were using i16 or larger, or using floats. i16 would make
some things simpler (eq comparisons, hashing, ordering, etc), but again
limits the precision. Using floats allows for varying the level of
precision displayed in the UI without having to make any lib changes.

+++ android/app/src/main/java/com/oppzippy/openscq30/features/equalizer/visualization/EqualizerLine.kt
@@ -5,7 +5,7 @@ import android.graphics.Canvas
 import android.graphics.Paint
 import com.oppzippy.openscq30.lib.bindings.VolumeAdjustments
 
-class EqualizerLine(private val values: List<Byte>) {
+class EqualizerLine(private val values: List<Double>) {
     fun drawBitmap(
         bitmap: Bitmap,
         xOffset: Float = 0F,
@@ -52,7 +52,7 @@ class EqualizerLine(private val values: List<Byte>) {
             val x = normalizedX * widthWithoutPadding + padding
             val normalizedY = 1F - ((value - minVolume) / range.toFloat())
             val y = normalizedY * heightWithoutPadding + padding
-            Pair(x, y)
+            Pair(x, y.toFloat())
         }
         return points
     }
